fix: 대문자로 시작하는 디렉토리 이름 수정  (#63)
 왜 대문자로 만들었냐
