+++ backend/src/main/java/com/example/backend/domain/information/model/car/entity/InteriorColor.java
@@ -1,16 +1,18 @@
-package com.example.backend.domain.Information.model.option.entity;
+package com.example.backend.domain.information.model.car.entity;
 
 import com.example.backend.domain.global.model.BaseInfo;
 import org.springframework.data.annotation.Id;
+import org.springframework.data.jdbc.core.mapping.AggregateReference;
 import org.springframework.data.relational.core.mapping.Embedded;
 import org.springframework.data.relational.core.mapping.Table;
 
-@Table("POWERTRAIN")
-public class Powertrain {
+@Table("INTERIOR_COLOR")
+public class InteriorColor {
     @Id
     private Long id;
+    AggregateReference<ExteriorColor, Long> exteriorColorId;
+    private String iconSrc;
     @Embedded(onEmpty = Embedded.OnEmpty.USE_NULL)
     private BaseInfo baseInfo;
-    private String detail;
     private String comment;
 }
