+++ backend/src/main/java/com/example/backend/domain/information/model/car/entity/Model.java
@@ -1,8 +1,8 @@
-package com.example.backend.domain.Information.model.car.entity;
+package com.example.backend.domain.information.model.car.entity;
 
-import com.example.backend.domain.Information.model.option.entity.Bodytype;
-import com.example.backend.domain.Information.model.option.entity.DrivingSystem;
-import com.example.backend.domain.Information.model.option.entity.Powertrain;
+import com.example.backend.domain.information.model.option.entity.Bodytype;
+import com.example.backend.domain.information.model.option.entity.DrivingSystem;
+import com.example.backend.domain.information.model.option.entity.Powertrain;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.jdbc.core.mapping.AggregateReference;
 import org.springframework.data.relational.core.mapping.Table;
