+++ backend/src/main/java/com/example/backend/domain/information/model/car/entity/Trim.java
@@ -1,12 +1,11 @@
-package com.example.backend.domain.Information.model.option.entity;
+package com.example.backend.domain.information.model.car.entity;
 
 import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Table;
 
-@Table("PARTS")
-public class Parts {
+@Table("TRIM")
+public class Trim {
     @Id
     private Long id;
     private String name;
-    private String imageSrc;
 }
