+++ backend/src/main/java/com/example/backend/domain/information/model/option/entity/Parts.java
@@ -1,11 +1,12 @@
-package com.example.backend.domain.Information.model.car.entity;
+package com.example.backend.domain.information.model.option.entity;
 
 import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Table;
 
-@Table("TRIM")
-public class Trim {
+@Table("PARTS")
+public class Parts {
     @Id
     private Long id;
     private String name;
+    private String imageSrc;
 }
