+++ backend/src/main/java/com/example/backend/domain/information/model/option/entity/Powertrain.java
@@ -1,16 +1,16 @@
-package com.example.backend.domain.Information.model.car.entity;
+package com.example.backend.domain.information.model.option.entity;
 
 import com.example.backend.domain.global.model.BaseInfo;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Embedded;
 import org.springframework.data.relational.core.mapping.Table;
 
-@Table("EXTERIOR_COLOR")
-public class ExteriorColor {
+@Table("POWERTRAIN")
+public class Powertrain {
     @Id
     private Long id;
-    private String colorCode;
-    @Embedded(onEmpty = Embedded.OnEmpty.USE_EMPTY)
+    @Embedded(onEmpty = Embedded.OnEmpty.USE_NULL)
     private BaseInfo baseInfo;
+    private String detail;
     private String comment;
 }
